refactor: missed blocks refactor (#1366)
