+++ app/Http/Controllers/Inertia/ValidatorsController.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers\Inertia;
 
-use App\DTO\Inertia\ForgingStats as ForgingStatsDTO;
 use App\Enums\SortDirection;
+use App\Http\Controllers\Inertia\Concerns\MissedBlocksTab;
 use App\Http\Controllers\Inertia\Concerns\RecentVotesTab;
 use App\Http\Controllers\Inertia\Concerns\ValidatorsTab;
 use App\Http\Controllers\Inertia\Concerns\WithPagination;
@@ -13,7 +13,6 @@ use App\Models\ForgingStats;
 use App\Services\Cache\NetworkCache;
 use App\Services\Cache\ValidatorCache;
 use ARKEcosystem\Foundation\UserInterface\UI;
-use Illuminate\Pagination\AbstractPaginator;
 use Inertia\Inertia;
 use Inertia\Response;
 
@@ -21,6 +20,7 @@ final class ValidatorsController
 {
     use ValidatorsTab;
     use RecentVotesTab;
+    use MissedBlocksTab;
     use WithPagination;
 
     public const FILTERS = [
@@ -111,46 +111,5 @@ final class ValidatorsController
     private function sortKey(string $name = 'default'): string
     {
         return 'rank';
-    }
-
-    private function getMissedBlocks(): AbstractPaginator
-    {
-        if (config('database.default') === 'sqlite') {
-            return ForgingStats::with('validator')
-                ->orderByDesc('timestamp')
-                ->whereNotNull('missed_height')
-                ->paginate($this->perPage(), page: $this->page())
-                ->through(fn (ForgingStats $voter) => ForgingStatsDTO::fromModel($voter));
-        }
-
-        // TODO: Re-implement sorting once the UI supports it - https://app.clickup.com/t/86dypp5jv
-        //       Look at \App\Http\Livewire\Validators\Concerns\MissedBlocksTab for reference.
-        $missedBlocksSortKey = 'height';
-        $sortDirection       = SortDirection::DESC;
-
-        return ForgingStats::query()
-            ->with('validator')
-            // @phpstan-ignore-next-line
-            ->when($missedBlocksSortKey === 'height', fn ($query) => $query->sortByHeight($sortDirection))
-            // @phpstan-ignore-next-line
-            ->when($missedBlocksSortKey === 'age', fn ($query) => $query->sortByAge($sortDirection))
-            // @phpstan-ignore-next-line
-            ->when($missedBlocksSortKey === 'name', fn ($query) => $query->sortByUsername($sortDirection))
-            // @phpstan-ignore-next-line
-            ->when($missedBlocksSortKey === 'votes' || $missedBlocksSortKey === 'percentage_votes', fn ($query) => $query->sortByVoteCount($sortDirection))
-            // @phpstan-ignore-next-line
-            ->when($missedBlocksSortKey === 'no_of_voters', fn ($query) => $query->sortByNumberOfVoters($sortDirection))
-            ->whereNotNull('missed_height')
-            ->paginate($this->perPage(), page: $this->page())
-            ->through(fn (ForgingStats $voter) => ForgingStatsDTO::fromModel($voter));
-    }
-
-    private function getMissedBlocksNoResultsMessageProperty(int $count): null|string
-    {
-        if ($count === 0) {
-            return trans('tables.missed-blocks.no_results');
-        }
-
-        return null;
     }
 }
